pyproject.toml
[project]
name = "forecast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-cloud-bigquery-storage>=2.32.0",
    "ipykernel>=6.29.5",
    "lightgbm>=4.6.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pandas-gbq>=0.29.2",
    "scikit-learn>=1.7.0",
    "tqdm>=4.67.1",
]
